sync-firestore/src/requirements.txt
google-cloud-firestore==2.11.1
cloud-sql-python-connector[pg8000]
sqlalchemy==2.0.0